Update gate-latch-monitor.py
Since the doorbell rang twice, once during the command for doorbell_is_on and once for doorbell_is_off I've added the if automation hat.input.one.read() == 1 to the activation of the doorbell on relay one. After this the bell works properly. It doesn't matter how long the switch is pressed it just rings for a second and waits until it is pressed again.